Домашние задания к уроку №8 Refactor  folders
