+++ app/src/main/java/ru/dw/gbkotlinweather/data/api_yandex/retrofit/WeatherApi.kt
@@ -1,10 +1,10 @@
-package ru.dw.gbkotlinweather.repository.api_yandex.retrofit
+package ru.dw.gbkotlinweather.data.api_yandex.retrofit
 
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Header
 import retrofit2.http.Query
-import ru.dw.gbkotlinweather.repository.api_yandex.WeatherDTO
+import ru.dw.gbkotlinweather.data.api_yandex.WeatherDTO
 import ru.dw.gbkotlinweather.utils.*
 
 interface WeatherApi {
