+++ app/src/main/java/ru/dw/gbkotlinweather/data/room/HistoryEntity.kt
@@ -1,4 +1,4 @@
-package ru.dw.gbkotlinweather.repository.local.room
+package ru.dw.gbkotlinweather.data.room
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
