+++ app/src/main/java/ru/dw/gbkotlinweather/data/room/MyDB.kt
@@ -1,4 +1,4 @@
-package ru.dw.gbkotlinweather.repository.local.room
+package ru.dw.gbkotlinweather.data.room
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
