+++ app/src/main/java/ru/dw/gbkotlinweather/repository/DetailsRepository.kt
@@ -1,7 +1,7 @@
 package ru.dw.gbkotlinweather.repository
 
 import ru.dw.gbkotlinweather.model.City
-import ru.dw.gbkotlinweather.view.viewmodel.DetailsViewModel
+import ru.dw.gbkotlinweather.view.details.DetailsViewModel
 
 fun interface DetailsRepository {
     fun getWeatherDetails(city: City, callbackDetailsWeather: DetailsViewModel.CallbackDetails)
