+++ app/src/main/java/ru/dw/gbkotlinweather/repository/DetailsRepositoryAdd.kt
@@ -1,8 +1,8 @@
 package ru.dw.gbkotlinweather.repository
 
 import ru.dw.gbkotlinweather.model.Weather
-import ru.dw.gbkotlinweather.view.viewmodel.DetailsViewModel
+import ru.dw.gbkotlinweather.view.details.DetailsViewModel
 
 interface DetailsRepositoryAdd {
-    fun addWeather(weather: Weather, callbackDetails:DetailsViewModel.CallbackDetails)
+    fun addWeather(weather: Weather, callbackDetails: DetailsViewModel.CallbackDetails)
 }
