+++ app/src/main/java/ru/dw/gbkotlinweather/repository/DetailsRepositoryForAll.kt
@@ -1,6 +1,6 @@
 package ru.dw.gbkotlinweather.repository
 
-import ru.dw.gbkotlinweather.view.viewmodel.DetailsViewModel
+import ru.dw.gbkotlinweather.view.details.DetailsViewModel
 
 fun interface DetailsRepositoryAll {
     fun getAllWeatherDetails(callbackWeather: DetailsViewModel.CallbackResponseForAll)
