+++ app/src/main/java/ru/dw/gbkotlinweather/repository/RepositoryImpl.kt
@@ -1,7 +1,7 @@
 package ru.dw.gbkotlinweather.repository
 
-import ru.dw.gbkotlinweather.repository.local.DataLocalListCity
-import ru.dw.gbkotlinweather.view.viewmodel.state.ListState
+import ru.dw.gbkotlinweather.data.local.DataLocalListCity
+import ru.dw.gbkotlinweather.view.state.ListState
 
 class RepositoryImpl : RepositoryListCity {
     private val data:RepositoryListCity = DataLocalListCity()
