+++ app/src/main/java/ru/dw/gbkotlinweather/view/details/DetailsViewModel.kt
@@ -1,4 +1,4 @@
-package ru.dw.gbkotlinweather.view.viewmodel
+package ru.dw.gbkotlinweather.view.details
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
@@ -6,10 +6,9 @@ import ru.dw.gbkotlinweather.MyApp
 import ru.dw.gbkotlinweather.model.City
 import ru.dw.gbkotlinweather.model.Weather
 import ru.dw.gbkotlinweather.repository.DetailsRepository
-import ru.dw.gbkotlinweather.repository.DetailsRepositoryRoomImpl
-import ru.dw.gbkotlinweather.repository.api_yandex.retrofit.DetailsRepositoryRetrofit
-import ru.dw.gbkotlinweather.utils.convertWeatherToEntity
-import ru.dw.gbkotlinweather.view.viewmodel.state.DetailsState
+import ru.dw.gbkotlinweather.repository.DetailsRepositoryImpl
+import ru.dw.gbkotlinweather.data.api_yandex.retrofit.RetrofitDetailsHelper
+import ru.dw.gbkotlinweather.view.state.DetailsState
 
 
 class DetailsViewModel(
@@ -24,9 +23,9 @@ class DetailsViewModel(
 
     fun upDataWeather(city: City) {
         repository = if (MyApp.getPref().getIsInternet()){
-            DetailsRepositoryRetrofit
+            RetrofitDetailsHelper
         }else {
-            DetailsRepositoryRoomImpl()
+            DetailsRepositoryImpl()
         }
 
         repository.getWeatherDetails(city,object : CallbackDetails {
